requirements.txt
jacinle
torch==2.1.0
torchvision==0.16.0
sapien==2.2.2
mani_skill2==0.5.3
mplib==0.2.0a0

diffusers==0.28.0
faiss_cpu==1.8.0
imageio==2.34.0
loguru==0.7.2
matplotlib==3.8.3
networkx==3.2.1
numpy==1.23.5
open3d==0.18.0
opencv_python_headless==4.10.0.84
packaging==24.1
pandas==2.2.3:q
Pillow==9.5.0
pybullet==3.2.6
scipy==1.14.1
tabulate==0.9.0
timm==0.6.11
tqdm==4.66.2
trimesh==4.2.0
